refactoring gallery pt 1/2
